service-telegram-bot/requirements.txt
python-telegram-bot==13.15
requests==2.26.0
python-dotenv==0.19.2
fastapi==0.95.0
uvicorn==0.21.1
